@@ -16,21 +16,23 @@
   "author": "freeman-lab",
   "license": "MIT",
   "dependencies": {
-    "@carbonplan/colormaps": "^3.0.2",
-    "@carbonplan/components": "^11.3.0",
-    "@carbonplan/minimaps": "^1.1.0",
-    "@carbonplan/prism": "^1.0.4",
-    "@carbonplan/theme": "^7.0.0",
+    "@carbonplan/colormaps": "^4.0.0",
+    "@carbonplan/components": "^12.0.0",
+    "@carbonplan/minimaps": "^2.1.0",
+    "@carbonplan/prism": "^2.0.0",
+    "@carbonplan/theme": "^8.0.0",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
-    "@mdx-js/loader": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@next/mdx": "^11.1.2",
-    "@theme-ui/color": "^0.13.1",
+    "@mdx-js/loader": "^2.1.5",
+    "@mdx-js/react": "^2.1.5",
+    "@next/mdx": "^12.3.1",
+    "@theme-ui/color": "^0.15.3",
+    "@theme-ui/mdx": "^0.15.3",
     "next": "^12.0.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "theme-ui": "^0.13.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "remark-mdx-code-meta": "^2.0.0",
+    "theme-ui": "^0.15.3"
   },
   "devDependencies": {
     "prettier": "^2.2.1"
