@@ -1,9 +1,16 @@
-const path = require('path')
-const withMDX = require('@next/mdx')({
+import path from 'path'
+import remarkMdxCodeMeta from 'remark-mdx-code-meta'
+import nextMDX from '@next/mdx'
+import * as url from 'url'
+
+const withMDX = nextMDX({
   extension: /\.mdx?$/,
+  options: { remarkPlugins: [remarkMdxCodeMeta], format: 'mdx' },
 })
 
-module.exports = withMDX({
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url))
+
+export default withMDX({
   pageExtensions: ['js', 'jsx', 'mdx', 'md'],
   webpack: (config, options) => {
     if (options.isServer) {
